+++ R/utils.R
@@ -51,16 +51,17 @@ ll_probit <- function(object) {
 }
 
 r2 <- function(object) {
-  z <- get_z(object)
-  y <- get_y(object)
-  RS <- residuals(object)**2
-  TS <- (y - mean(y))**2
+  z <- OPSR:::get_z(object)
+  y <- OPSR:::get_y(object)
+  RS <- residuals(object)^2
+  TS <- (y - mean(y))^2
   R2o <- unlist(lapply(seq_len(object$nReg), function(i) {
     RSS <- sum(RS[z == i])
-    TSS <- sum(TS[z == i])
-    1 - RSS / TSS
+    yo <- y[z == i]
+    TSS <- sum((yo - mean(yo))^2)
+    1 - RSS/TSS
   }))
-  R2total <- 1 - sum(RS) / sum(TS)
+  R2total <- 1 - sum(RS)/sum(TS)
   R2 <- c(R2total, R2o)
   names(R2) <- c("Total", paste0("o", 1:object$nReg))
   R2