+++ NEWS.md
@@ -1,9 +1,10 @@
 # OPSR (development version)
 
+* Added GOF indicators to print.summary.opsr()
 * Added timeuse data
 * R2 for each regime in texreg functions.
 * Nicer summary printing and setting fixed values in null model to NA.
-* Fixed issues 7 - 9.
+* Fixed issues 7 - 10.
 
 # OPSR 0.1.2