+++ NEWS.md
@@ -4,7 +4,9 @@
 * Added timeuse data
 * R2 for each regime in texreg functions.
 * Nicer summary printing and setting fixed values in null model to NA.
-* Fixed issues 7 - 10.
+* Added `.loglik` arg to `opsr()` to return log-likelihood vector.
+* Detects and errors on singularity issues.
+* Fixed issues 7 - 12.
 
 # OPSR 0.1.2